LayoutTests/accessibility/mac/webarea-size-equals-content-size-expected.txt


text
text
text
text
text
text
text
text
This tests that an ARIA menu bar is exposed as a menu bar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.role is 'AXRole: AXWebArea'
PASS accessibilityController.focusedElement.parentElement().role is 'AXRole: AXScrollArea'
PASS webAreaWidth > scrollWidth is true
PASS webAreaHeight > scrollHeight is true
PASS successfullyParsed is true

TEST COMPLETE
